multiprocess returns same posterior val and is faster
